Enhance and refactor loading indicator
